# openapi/transcendence.v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Transcendence API",
    "description": "League analytics API with app API-key and user JWT authentication.",
    "version": "v1"
  },
  "paths": {
...
    "/api/admin/pro-summoners": {
      "get": {
        "tags": [
          "ProSummoners"
        ],
        "parameters": [
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackedProSummonerDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackedProSummonerDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackedProSummonerDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          { }
        ]
      },
      "post": {
        "tags": [
          "ProSummoners"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTrackedProSummonerRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTrackedProSummonerRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTrackedProSummonerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProSummonerDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProSummonerDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProSummonerDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/admin/pro-summoners/{id}": {
      "get": {
        "tags": [
          "ProSummoners"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProSummonerDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProSummonerDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProSummonerDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          { }
        ]
      },
      "put": {
        "tags": [
          "ProSummoners"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTrackedProSummonerRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTrackedProSummonerRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTrackedProSummonerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProSummonerDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProSummonerDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedProSummonerDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          { }
        ]
      },
      "delete": {
        "tags": [
          "ProSummoners"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/summoners/{region}/{name}/{tag}": {
      "get": {
        "tags": [
          "Summoners"
        ],
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SummonerProfileResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummonerProfileResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummonerProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SummonerAcceptedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummonerAcceptedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummonerAcceptedResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataAgeMetadata": {
        "type": "object",
        "properties": {
          "fetchedAt": {
            "type": "string",
            "format": "date-time"
          },
          "age": {
            "type": "string",
            "format": "date-span",
            "readOnly": true
          },
          "ageDescription": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProfileChampionStat": {
        "type": "object",
        "properties": {
          "championId": {
            "type": "integer",
            "format": "int32"
          },
          "championName": {
            "type": "string",
            "nullable": true
          },
          "games": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "double"
          },
          "kdaRatio": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ProfileOverviewStats": {
        "type": "object",
        "properties": {
          "totalMatches": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "double"
          },
          "avgKills": {
            "type": "number",
            "format": "double"
          },
          "avgDeaths": {
            "type": "number",
            "format": "double"
          },
          "avgAssists": {
            "type": "number",
            "format": "double"
          },
          "kdaRatio": {
            "type": "number",
            "format": "double"
          },
          "avgCsPerMin": {
            "type": "number",
            "format": "double"
          },
          "avgVisionScore": {
            "type": "number",
            "format": "double"
          },
          "avgDamageToChamps": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ProfileRecentMatch": {
        "type": "object",
        "properties": {
          "matchId": {
            "type": "string",
            "nullable": true
          },
          "matchDate": {
            "type": "integer",
            "format": "int64"
          },
          "queueType": {
            "type": "string",
            "nullable": true
          },
          "win": {
            "type": "boolean"
          },
          "championId": {
            "type": "integer",
            "format": "int32"
          },
          "championName": {
            "type": "string",
            "nullable": true
          },
          "kills": {
            "type": "integer",
            "format": "int32"
          },
          "deaths": {
            "type": "integer",
            "format": "int32"
          },
          "assists": {
            "type": "integer",
            "format": "int32"
          },
          "csPerMin": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "RankInfo": {
        "type": "object",
        "properties": {
          "tier": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "type": "string",
            "nullable": true
          },
          "leaguePoints": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SummonerAcceptedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "poll": {
            "type": "string",
            "nullable": true
          },
          "retryAfterSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SummonerProfileResponse": {
        "type": "object",
        "properties": {
          "summonerId": {
            "type": "string",
            "format": "uuid"
          },
          "puuid": {
            "type": "string",
            "nullable": true
          },
          "gameName": {
            "type": "string",
            "nullable": true
          },
          "tagLine": {
            "type": "string",
            "nullable": true
          },
          "summonerLevel": {
            "type": "integer",
            "format": "int32"
          },
          "profileIconId": {
            "type": "integer",
            "format": "int32"
          },
          "soloRank": {
            "$ref": "#/components/schemas/RankInfo"
          },
          "flexRank": {
            "$ref": "#/components/schemas/RankInfo"
          },
          "overviewStats": {
            "$ref": "#/components/schemas/ProfileOverviewStats"
          },
          "topChampions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileChampionStat"
            },
            "nullable": true
          },
          "recentMatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileRecentMatch"
            },
            "nullable": true
          },
          "profileAge": {
            "$ref": "#/components/schemas/DataAgeMetadata"
          },
          "rankAge": {
            "$ref": "#/components/schemas/DataAgeMetadata"
          },
          "statsAge": {
            "$ref": "#/components/schemas/DataAgeMetadata"
          }
        },
        "additionalProperties": false
      },
...
      "TrackedProSummonerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "puuid": {
            "type": "string",
            "nullable": true
          },
          "platformRegion": {
            "type": "string",
            "nullable": true
          },
          "gameName": {
            "type": "string",
            "nullable": true
          },
          "tagLine": {
            "type": "string",
            "nullable": true
          },
          "proName": {
            "type": "string",
            "nullable": true
          },
          "teamName": {
            "type": "string",
            "nullable": true
          },
          "isPro": {
            "type": "boolean"
          },
          "isHighEloOtp": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "UpsertTrackedProSummonerRequest": {
        "type": "object",
        "properties": {
          "puuid": {
            "type": "string",
            "nullable": true
          },
          "platformRegion": {
            "type": "string",
            "nullable": true
          },
          "gameName": {
            "type": "string",
            "nullable": true
          },
          "tagLine": {
            "type": "string",
            "nullable": true
          },
          "proName": {
            "type": "string",
            "nullable": true
          },
          "teamName": {
            "type": "string",
            "nullable": true
          },
          "isPro": {
            "type": "boolean"
          },
          "isHighEloOtp": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "App authentication key for operational endpoints.",
        "name": "X-API-Key",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "User JWT access token from /api/auth/login.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}